backend/requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
requests
pydantic
pyautogui==0.9.54
pygetwindow==0.0.9
pyperclip==1.8.2
networkx==3.1
numpy==1.24.3
plotly==5.18.0